Refactor: add package and move class
