+++ src/test/java/com/cheolhyeonpark/step4/domain/CarTest.java
@@ -1,4 +1,4 @@
-package com.cheolhyeonpark.step4;
+package com.cheolhyeonpark.step4.domain;
 
 import com.cheolhyeonpark.step4.domain.Car;
 import org.junit.jupiter.api.DisplayName;
