+++ src/test/java/com/cheolhyeonpark/step4/domain/DiceTest.java
@@ -1,4 +1,4 @@
-package com.cheolhyeonpark.step4;
+package com.cheolhyeonpark.step4.domain;
 
 import com.cheolhyeonpark.step4.domain.Dice;
 import org.junit.jupiter.api.DisplayName;
