+++ src/test/java/com/cheolhyeonpark/step4/domain/RacingGameTest.java
@@ -1,7 +1,5 @@
-package com.cheolhyeonpark.step4;
+package com.cheolhyeonpark.step4.domain;
 
-import com.cheolhyeonpark.step4.domain.Car;
-import com.cheolhyeonpark.step4.domain.RacingGame;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
